+++ .rsink/rsink.sh
@@ -37,14 +37,13 @@ warn() { # show error, don't exit
 
 # Usage: usage
 #---------------------------------------------------------
-# Display usage. Uses constants from constants.sh
+# Display usage.
 #---------------------------------------------------------
 usage() {
     echo -e "$VERSION\n"
     echo -e "./rsink.sh <options>"
     echo -e "\t-d or --dry-run     # Dry run (Don't run rsync)"
     echo -e "\t-h or --help        # Display help"
-    echo -e "\t-p or --pushover    # Send Pushover.net notification"
     echo -e "\t-s or --silent      # Silent output"
     echo -e "\t-v or --version     # Displays version\n"
     echo "Config file: $CONFIG_FILE"
@@ -53,12 +52,14 @@ usage() {
 
 # Usage: detect_x $1 $2
 #---------------------------------------------------------
-# Error has occurred. Warn user.
+# Determine if files, and/or directories exist. Check if
+# volumes are mounted.
 #
 # Inputs
 #   $1  Type
 #       * "f" File
 #       * "d" Directory
+#       * "L" Symbolic link
 #       * "x" File or directory
 #       * "m" Mounted volume
 #   $2  File/folder name
@@ -73,12 +74,16 @@ detect_x() {
             if [ ! -d "$2" ]; then
                 fail "$2" "Directory does not exist."
             fi ;;
+        "L"*) # Symbolic link
+            if [ ! -L "$2" ]; then
+                warn "$2 is not a symbolic link. Proceeding."
+            fi ;;
         "x"*) # Either file or directory
             if [ ! -f "$2" ] && [ ! -d "$2" ]; then
                 fail "$2" "File or directory does not exist."
             fi ;;
         "m"*) # Mounted volume
-            if [ "$(mount | grep -c $2)" -ne 1 ]; then
+            if ! mount | grep "$2" > /dev/null; then
                 warn "$2 not mounted. Skipping."
                 skip=1
             else
@@ -89,33 +94,26 @@ detect_x() {
 
 # Usage: profile $1
 #---------------------------------------------------------
-# Parse profile to build rsync command's options.
+# Parse profile to build rsync command's options in
+# $args array.
 #
 # Inputs
 #   $1  profile file
 #---------------------------------------------------------
-profile() { # Parse profile file
-    detect_x "f" "$PROFILES_DIRECTORY/$1" # Detect profile
+profile(){
+    while read line; do
+        arg=${line%#*} # Disregard comment
+        arg="${arg%%*( )}" # Trim trailing whitespace
 
-    local first=1 # First single character option (add -)
-    while read -r line || [ -n "$line" ]; do # Parse profile file
-        local option=$(printf "%s" "$line" | awk '{print $1;}') # Get first token of $line
-
-        if [ "$(printf "%s" "$option" | head -c 1)" != "#" ] && [ "$option" != "" ]; then # Not comment or empty line
-            if [ ${#option} -eq 1 ]; then # Single character options (add -)
-                if [ $first -eq 1 ]; then
-                    first=0
-                    printf " -%s" "$option"
-                else
-                    printf "%s" "$option"
-                fi
-            else # Long options (add --)
-                first=1
-                printf " --%s" "$option"
+        # Only consider non-comment and non-empty lines
+        if [[ $(printf "%s" "$line" | head -c 1) != "#" && -n "$line" ]]; then
+            if [ "${#arg}" -eq 1 ]; then
+                args=("${args[@]}" "-$arg")
+            else
+                args=("${args[@]}" "--$arg")
             fi
         fi
-    done < "$PROFILES_DIRECTORY/$1"
-    printf " --log-file=\"%s\"" "$LOG_FILE"
+    done < "$1"
 }
 
 # Usage: isBackup $1
@@ -128,10 +126,13 @@ profile() { # Parse profile file
 isBackup() {
     backup=0
     current_version=""
-    local line=$(grep "link-dest" "$PROFILES_DIRECTORY/$1")
+    local line="$(grep "link-dest" "$1")"
     if [ "$line" != "" ]; then
         backup=1
-        current_version=$(printf "%s" "$line" | cut -d'=' -f2)
+        current_version=${line#"link-dest="}
+        current_version=${current_version#"'"}
+        current_version=${current_version%"'"}
+        detect_x "L" "$current_version"
     fi
 }
 
@@ -143,93 +144,84 @@ main() {
     detect_x "f" "$CONFIG_FILE"
     detect_x "d" "$PROFILES_DIRECTORY"
 
-    while read line || [ -n "$line" ]; do # Parse config file
-        local token=0
-        local cmd="rsync"
-        backup=0
-        skip=0
-
-        for p in $line; do # Parse tokens in config file
-            (( token++ ))
-
-            if [ $token -eq 1 ]; then # Profile
-                local profile=$(printf "%s" "$p" | awk '{print $1;}') # Get first word of $line
-                cmd="$cmd$(profile $profile)"
-            elif [ $token -eq 2 ]; then # Source
-                if [ "$(printf "%s" "$p" | head -c 1)" = "~" ]; then # ~/source/folder
-                    p=$(printf "%s" "$p" | cut -d "~" -f 2)
-                    p="$HOME$p" # /Users/user/source/folder or /home/source/folder
-                fi
-                local src=$p
-                detect_x "x" "$p"
-                cmd="$cmd $p"
-            elif [ $token -eq 3 ]; then # Destination volume
-                local dest=$p
-                detect_x "m" "$p"
-                if [ $skip -eq 1 ]; then
-                    break;
-                fi
-                cmd="$cmd $p"
-            elif [ $token -eq 4 ]; then # Destination folder
-                if [ "$p" != "." ]; then
-                    isBackup "$profile"
-                    if [ $backup -ge 1 ]; then
-                        local ddate=$(date +"-%m-%d-%Y @ %H-%M-%S") # Dash and date
-                        dest="$dest/$p$ddate"
-                        cmd="$cmd/$p$ddate"
-                    else
-                        detect_x "d" "$dest/$p"
-                        cmd="$cmd/$p"
-                    fi
-                fi
-            elif [ $token -ge 5 ]; then # Excludes
-                detect_x "x" "$src/$p"
-                cmd="$cmd --exclude='$p'"
+    backup=0
+    skip=0
+    token=1
+    while read -r line; do # Parse config file
+    if [[ "$(printf "%s" "$line" | awk '{print $1;}')" != "#" ]]; then
+        if [[ "$line" == "" ]]; then
+            echo "> rsync" "${args[@]}" "\"$src\"" "\"$dest\"" "${excludes[@]}" --log-file="'$LOG_FILE'"
+            if [[ "$dry" -ne 1 && "$skip" -ne 1 ]]; then
+                rsync "${args[@]}" "$src" "$dest" "${excludes[@]}" --log-file="$LOG_FILE" > /dev/null
             fi
-        done
-        token=0
-
-        # Run rsync
-        echo -e "$cmd\n"
-        if [[ $dry -ne 1 && $skip -ne 1 ]]; then
-            eval "$cmd" > /dev/null
-        fi
-        code=$?
-
-        if [ $dry -ne 1 ]; then
-            # Symlink new backup to link-dest path in profile
-            if [ $backup -eq 1 ]; then
-                rm -rf "$current_version"
-                ln -s "$dest $current_version"
-            fi
+            code=$?
+            args=()
+            excludes=()
+            token=0
 
             # rsync error checking
-            if [ $code -ne 0 ]; then
+            if [ "$code" -ne 0 ]; then
                 warn "rsync code: $code\nrsync completed with errors."
             fi
 
-            # Log file errors
-            if [ $skip -ne 1 ]; then
-                if grep -cq "No space left on device (28)\|Result too large (34)" "$LOG_FILE"; then
-                    warn "Not enough space on $dest"
+            if [[ "$dry" -ne 1 && "$skip" -ne 1 ]]; then
+                if [ "$backup" -eq 1 ]; then
+                    # Symlink new backup to link-dest path in profile
+                    rm -rf "$current_version"
+                    ln -s "$dest" "$current_version"
+                fi
+
+                # Log file errors
+                if [ -n "$(grep "No space left on device (28)\|Result too large (34)" "$LOG_FILE")" ]; then
+                    warn "Not enough space on $volume"
                 fi
                 rm "$LOG_FILE"
             fi
-        fi
 
-        skip=0
-        backup=0
-        current_version=""
+            skip=0
+            backup=0
+            current_version=""
+        elif [ "$token" -eq 1 ]; then # Profile
+            detect_x "f" "$PROFILES_DIRECTORY/$line"
+            profile="$PROFILES_DIRECTORY/$line"
+            profile "$profile"
+        elif [ "$token" -eq 2 ]; then # Source
+            detect_x "x" "$line"
+            src="$line"
+        elif [ "$token" -eq 3 ]; then # Destination volume
+            detect_x "m" "$line"
+            volume="$line"
+            dest="$line"
+        elif [[ "$token" -eq 4 && "$skip" -ne 1 ]]; then # Destination folder
+            if [ "$line" != "." ]; then
+                isBackup "$profile"
+                if [ $backup -ge 1 ]; then
+                    local ddate=$(date +"-%m-%d-%Y@%H-%M-%S") # Dash and date
+                    dest="$dest/$line$ddate"
+                else
+                    detect_x "d" "$dest/$line"
+                    dest="$dest/$line"
+                fi
+            fi
+        elif [[ "$token" -ge 5 && "$skip" -ne 1 ]]; then # Excludes
+            detect_x "x" "$src/$line"
+            excludes=("${excludes[@]}" "--exclude=\"$line\"")
+        fi
+        (( token++ ))
+    fi
     done < "$CONFIG_FILE"
 }
 
-set -e # exit if any program exists with exit status > 0
+set -e # Exit if any program exists with exit status > 0
+shopt -s extglob # Extended pattern matching
 
-source ~/.rsink/"constants.sh"
+VERSION="rsink 0.2+dev"
+CONFIG_FILE="config"
+PROFILES_DIRECTORY="profiles"
+LOG_FILE="log"
 
 # Options parsing
 dry=0
-pushover=0
 silent=0
 
 while : ; do
@@ -240,9 +232,6 @@ while : ; do
         -h | --help | -\?)
             usage
             exit 0 ;;
-        -p | --pushover)
-            pushover=1
-            shift ;;
         -s | --silent)
             silent=1
             shift ;;
@@ -264,10 +253,5 @@ if [ $silent -eq 1 ]; then
 fi
 
 main
-
-# Send pushover notification
-if [ $pushover -eq 1 ]; then
-    ./"$TOOLS_DIRECTORY"/pushover.sh
-fi
 
 exit 0